Move around some code in completion.py
Because we have so little code now for the completion logic, it doesn't
make much sense anymore to keep around relevant code in various files.

- Move all the stuff in plugin/core/completion.py to plugin/completion.py
- Move all the stuff in restore_lines.py to completion.py
- Remove the tests in test_completion_core.py, I am confident in the
  functionality working as intended
- Streamline the tests in test_completion.py
